--- src/main/java/dat3/cars/repository/MemberRepository.java
@@ -1,8 +1,0 @@
-package dat3.cars.repository;
-
-import dat3.cars.entity.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member,String> {
-  boolean existsByEmail(String email);
-}
