+++ src/main/java/dat3/rename_me/api/info.txt
@@ -1,0 +1,1 @@
+API (Controller) classes go here
