+++ src/main/java/dat3/rename_me/configuration/SetupDevUsers.java
@@ -1,0 +1,47 @@
+package dat3.rename_me.configuration;
+
+import dat3.security.entity.Role;
+import dat3.security.entity.UserWithRoles;
+import org.springframework.boot.ApplicationArguments;
+import org.springframework.boot.ApplicationRunner;
+import org.springframework.stereotype.Controller;
+import dat3.security.repository.UserWithRolesRepository;
+
+@Controller
+public class SetupDevUsers implements ApplicationRunner {
+
+    UserWithRolesRepository userWithRolesRepository;
+    String passwordUsedByAll;
+
+    public SetupDevUsers(UserWithRolesRepository userWithRolesRepository) {
+        this.userWithRolesRepository = userWithRolesRepository;
+        passwordUsedByAll = "test12";
+    }
+
+    @Override
+    public void run(ApplicationArguments args) {
+        setupUserWithRoleUsers();
+    }
+
+    /*****************************************************************************************
+     NEVER  COMMIT/PUSH CODE WITH DEFAULT CREDENTIALS FOR REAL
+     iT'S ONE OF THE TOP SECURITY FLAWS YOU CAN DO
+     *****************************************************************************************/
+    private void setupUserWithRoleUsers() {
+        System.out.println("******************************************************************************");
+        System.out.println("******* NEVER  COMMIT/PUSH CODE WITH DEFAULT CREDENTIALS FOR REAL ************");
+        System.out.println("******* REMOVE THIS BEFORE DEPLOYMENT, AND SETUP DEFAULT USERS DIRECTLY  *****");
+        System.out.println("**** ** ON YOUR REMOTE DATABASE                 ******************************");
+        System.out.println("******************************************************************************");
+        UserWithRoles user1 = new UserWithRoles("user1", passwordUsedByAll, "user1@a.dk");
+        UserWithRoles user2 = new UserWithRoles("user2", passwordUsedByAll, "user2@a.dk");
+        UserWithRoles user3 = new UserWithRoles("user3", passwordUsedByAll, "user3@a.dk");
+        user1.addRole(Role.USER);
+        user1.addRole(Role.ADMIN);
+        user2.addRole(Role.USER);
+        user3.addRole(Role.ADMIN);
+        userWithRolesRepository.save(user1);
+        userWithRolesRepository.save(user2);
+        userWithRolesRepository.save(user3);
+    }
+}
