+++ src/main/java/dat3/rename_me/dto/info.txt
@@ -1,0 +1,1 @@
+Add DTO (Data Transfer Object) classes here
