+++ src/main/java/dat3/rename_me/entity/info.txt
@@ -1,0 +1,1 @@
+Add Entity classes here
