+++ src/main/java/dat3/rename_me/service/info.txt
@@ -1,0 +1,1 @@
+Add Service Classes/Interfaces here
