+++ src/main/resources/application.properties
@@ -1,9 +1,8 @@
 spring.jpa.hibernate.ddl-auto=create
 
-
-spring.datasource.url=${JDBC_DATABASE_URL}
-spring.datasource.username=${JDBC_USERNAME}
-spring.datasource.password=${JDBC_PASSWORD}
+#spring.datasource.url=${JDBC_DATABASE_URL}
+#spring.datasource.username=${JDBC_USERNAME}
+#spring.datasource.password=${JDBC_PASSWORD}
 
 #JDBC_DATABASE_URL=jdbc:mysql://localhost:3306/cars_fall2022;JDBC_USERNAME=dev;JDBC_PASSWORD=ax2
